pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cornercrop"
version = "0.1.0"
description = "macOS Vision-powered corner watermark detector & cropper"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "CornerCrop contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
dependencies = [
    "Pillow>=9.0",
    "pyobjc-framework-Vision>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-CoreML>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-core>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
cornercrop = "cornercrop.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
